requirements.txt
SAM_2-1.0-py3-none-any.whl
Flask
matplotlib
opencv-python
vtracer 